@@ -19,8 +19,7 @@
   },
   "keywords": [
     "angular",
-    "angularjs",
-    "angular5",
+    "angular6",
     "ng",
     "ng2",
     "table",
@@ -35,37 +34,37 @@
   },
   "homepage": "https://github.com/mariuszfoltak/angular2-datatable#readme",
   "devDependencies": {
-    "@angular/core": "^5.0.0",
-    "@angular/compiler": "^5.0.0",
-    "@angular/common": "^5.0.0",
-    "@angular/platform-browser": "^5.0.0",
-    "@angular/platform-browser-dynamic": "^5.0.0",
-    "@angular/compiler-cli": "^5.0.0",
-    "core-js": "^2.5.3",
+    "@angular/core": "^6.0.0",
+    "@angular/compiler": "^6.0.0",
+    "@angular/common": "^6.0.0",
+    "@angular/platform-browser": "^6.0.0",
+    "@angular/platform-browser-dynamic": "^6.0.0",
+    "@angular/compiler-cli": "^6.0.0",
+    "core-js": "^2.5.5",
     "http-server": "^0.11.1",
     "jasmine-core": "^2.8.0",
-    "karma": "^2.0.0",
+    "karma": "^2.0.2",
     "karma-chrome-launcher": "^2.2.0",
-    "karma-coverage": "^1.1.1",
-    "karma-jasmine": "^1.1.1",
+    "karma-coverage": "^1.1.2",
+    "karma-jasmine": "^1.1.2",
     "karma-phantomjs-launcher": "^1.0.4",
     "phantomjs-prebuilt": "^2.1.7",
     "remap-istanbul": "^0.10.1",
-    "rxjs": "^5.4.3",
-    "systemjs": "^0.21.0",
-    "typescript": "2.6.x",
-    "zone.js": "^0.8.4",
+    "rxjs": "^6.0.0",
+    "systemjs": "^0.21.3",
+    "typescript": "2.7.x",
+    "zone.js": "^0.8.26",
     "rimraf": "^2.6.2",
-    "@types/lodash": "4.14.104",
-    "@types/jasmine": "^2.8.6"
+    "@types/lodash": "4.14.108",
+    "@types/jasmine": "^2.8.7"
   },
   "dependencies": {
-    "lodash": "^4.0.0"
+    "lodash": "^4.17.10"
   },
   "peerDependencies": {
-    "@angular/core": "^5.0.0",
-    "@angular/common": "^5.0.0",
-    "@angular/platform-browser": "^5.0.0",
-    "rxjs": "^5.4.3"
+    "@angular/core": "^6.0.0",
+    "@angular/common": "^6.0.0",
+    "@angular/platform-browser": "^6.0.0",
+    "rxjs": "^6.0.0"
   }
 }
