@@ -30,8 +30,13 @@
                 defaultExtension: 'js'
             },
             rxjs: {
+                main: 'index.js',
                 defaultExtension: 'js'
-            }
+            },
+            'rxjs/operators' : {
+                main: 'index.js',
+                defaultExtension: 'js'
+            },
         }
     });
 })(this);
